fix: boot-mode: Disregard RDR, as we have separate RX and TX lines
The original code we copied was for the H8/539F in a Mitsubishi ECM,
where the RX and TX lines were a single line (the K-line):

https://docs.google.com/spreadsheets/d/1Sq9vEf4kTzgJw3Uy3LirPGfeFqmQMafg_q5V_XRRS0s/edit#gid=0

Every character transmitted would end up being visible to the RX line
as well, so it'd end up in the Receive Data Register, and so we'd need
to manually clear the RDR after every transmit.

We don't need ANY of that this time -- we don't even need to set up
any of the receive configuration at all at this point -- so don't.