feat: boot-mode: pin SCI1 to 4800 baud, and document it
We select 4800 baud pretty arbitrarily, it's just what I remember
setting my scripts to when initially getting the program into boot
mode. We're in async mode as we don't have an external clock to align
serial timing.

This commit will also help you understand how to change the baud rate
later.

Walking through the documentation:

Table 12-3 of the HWM here[^1] tells us that for n=0, N=64 and
ø=10MHz, we'd get a bit-rate of 4800 baud with +0.16% error.

The definitions of n, N and ø are given in a legend right below those
tables here[^2].

- n is CKS1 * 2 + CKS0[^3]

- N is the value of the BRR[^4] taken as a uint8

- ø is the CPU clock frequency, in our case our crystal runs at 10MHz.

If we give a desired bit-rate B and CPU frequency ø, we can calculate
values for n and N in async mode using the formula next to the
aforementioned legend, here[^2].

[^1]:
https://downloads.laboratoryb.org/insight/documents/H8_538F_HWM.PDF#page=377

[^2]:
https://downloads.laboratoryb.org/insight/documents/H8_538F_HWM.PDF#page=379

[^3]:
https://downloads.laboratoryb.org/insight/documents/H8_538F_HWM.PDF#page=365

[^4]:
https://downloads.laboratoryb.org/insight/documents/H8_538F_HWM.PDF#page=375